Unified webpack; need more configuration
