--- src/lambdas/ListFiles/.babelrc
@@ -1,13 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/preset-env",
-      {
-        "targets": {
-          "node": "12"
-        }
-      }
-    ],
-    ["@babel/preset-typescript"]
-  ]
-}
