--- src/lambdas/ListFiles/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "listfiles",
-  "version": "1.0.0",
-  "description": "",
-  "main": "webpack.config.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack"
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@babel/core": "^7.12.3",
-    "@babel/preset-env": "^7.12.1",
-    "@babel/preset-typescript": "^7.12.1",
-    "babel-loader": "^8.1.0",
-    "cache-loader": "^4.1.0",
-    "fork-ts-checker-webpack-plugin": "^6.0.0",
-    "ts-loader": "^8.0.10",
-    "typescript": "^4.0.5",
-    "webpack": "^5.4.0",
-    "webpack-cli": "^4.2.0"
-  }
-}
