--- src/lambdas/ListFiles/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "es2015",
-    "target": "esnext",
-    "outDir": "./dist",
-    "strict": false,
-    "types": ["node"],
-    "resolveJsonModule": true,
-    "moduleResolution": "node",
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noImplicitReturns": true,
-    "noFallthroughCasesInSwitch": true,
-    "esModuleInterop": true,
-    "lib": ["ES2020.Promise"]
-  },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "**/*.test.ts"]
-}
